+++ server/include/Entities/Actions/BetAction.h
@@ -5,7 +5,7 @@
 
 class BetAction : public Action {
 public:
-    BetAction(string id, size_t amount) : Action(id, amount) {}
+    BetAction(string id, uint32_t amount) : Action(id, amount) {}
     ActionType getActionType() const override { return ActionType::BET; }
 };
