+++ server/include/Entities/Actions/RaiseAction.h
@@ -5,7 +5,7 @@
 
 class RaiseAction : public Action {
 public:
-    RaiseAction(string id, size_t amount) : Action(id, amount) {} 
+    RaiseAction(string id, uint32_t amount) : Action(id, amount) {} 
     ActionType getActionType() const override { return ActionType::RAISE; }
 };
