+++ server/include/Entities/PossibleActions/PossibleAllInBet.h
@@ -5,9 +5,9 @@
 
 class PossibleAllInBet : public PossibleAction {
 public:
-    PossibleAllInBet(size_t amount) : PossibleAction(amount) {}
+    PossibleAllInBet(uint32_t amount) : PossibleAction(amount) {}
     ActionType getActionType() const override { return ActionType::ALL_IN_BET; }
-    const size_t getSecondaryAmount() const override { return 0; }
+    const uint32_t getSecondaryAmount() const override { return 0; }
 };
 
 #endif
