+++ server/include/Entities/PossibleActions/PossibleCheck.h
@@ -7,7 +7,7 @@ class PossibleCheck : public PossibleAction {
 public:
     PossibleCheck() : PossibleAction(0) {}
     ActionType getActionType() const override { return ActionType::CHECK; }
-    const size_t getSecondaryAmount() const override { return 0; }
+    const uint32_t getSecondaryAmount() const override { return 0; }
 };
 
 #endif
