+++ server/src/Services/IOManager.cpp
@@ -14,14 +14,14 @@ void IOManager::displayPossibleActionsStdOut() {
     PrintUtil::printPossibleActionsForCurPlayer(gameData);
 }
 
-vector<pair<string, size_t>> IOManager::addPlayersStdIn() {
-    vector<pair<string, size_t>> newPlayers;
+vector<pair<string, uint32_t>> IOManager::addPlayersStdIn() {
+    vector<pair<string, uint32_t>> newPlayers;
     while (true) {
         cout << "(+) I/O Manager: Would you like to add a new player? ";
         if (!IOUtil::getYesOrNo()) break;
 
         string name = IOUtil::getPlayerName();
-        size_t chips = IOUtil::getPlayerChips(gameData.getBigBlind());
+        uint32_t chips = IOUtil::getPlayerChips(gameData.getBigBlind());
         newPlayers.push_back({name, chips});
     }
     cout << endl;
@@ -41,11 +41,11 @@ vector<string> IOManager::removePlayersStdIn() {
     return playersToRemove;
 }
 
-pair<ActionType, size_t> IOManager::getPlayerActionStdIn() {
+pair<ActionType, uint32_t> IOManager::getPlayerActionStdIn() {
     cout << "(+) I/O Manager Please input an action. E.g. ('bet 50' or 'fold'): ";
     auto possibleActions = gameData.getPossibleActions();
     ActionType type = IOUtil::getActionType(possibleActions);
-    size_t amount = IOUtil::getActionAmount(possibleActions, type);
+    uint32_t amount = IOUtil::getActionAmount(possibleActions, type);
     cout << endl;
     return {type, amount};
 }
